Move many MPS related autofermion tests to LegacyMPS
